+++ R/autoplot.sim_results.R
@@ -1,9 +1,17 @@
 #' Plot simulation results
 #' `autoplot()` method for `sim_results` objects
 #'
-#' @param object a `sim_results` object (from [analyze_sims()])
-#' @param statistic which column of `object` should be the y-axis?
+#' @param object a [data.frame]
+#' containing the columns expected
+#' for a `sim_results` object (from [analyze_sims()])
+#' @param statistic which column of `object` should be the y-axis
+#' @param x_var [character]: which column in `object` to use for the x-axis
 #' @param ... unused
+#' @param group_var [character]: which column in `object` to use for the
+#' `group` aesthetic in [ggplot2::aes()]
+#' @param color_var [character]: which column in `object` to use for the
+#' `color` aesthetic in [ggplot2::aes()]
+#'
 #' @returns a [ggplot2::ggplot]
 #' @export
 #'
@@ -11,14 +19,17 @@
 autoplot.sim_results <- function(
     object,
     statistic = "Empirical_SE",
+    x_var = "sample_size",
+    group_var = "lambda.sim",
+    color_var = group_var,
     ...) {
   object |>
     dplyr::mutate(lambda.sim = factor(.data$lambda.sim)) |>
     ggplot2::ggplot() +
     ggplot2::aes(
-      x = .data$sample_size,
-      group = .data$lambda.sim,
-      col = .data$lambda.sim,
+      x = .data[[x_var]],
+      group = .data[[group_var]],
+      col = .data[[color_var]],
       y = .data[[statistic]]
     ) +
     ggplot2::geom_point() +