+++ NEWS.md
@@ -19,6 +19,7 @@
 `autoplot.curve_params()` (#453)
 * Extended `sim_pop_data_multi()` to loop over multiple sample sizes (#444)
 * Added new functions `analyze_sims()` and `autoplot.sim_results()` (#444)
+* Extended `autoplot.sim_results()` to allow customization of x-axis, group, and color aesthetics via `x_var`, `group_var`, and `color_var` parameters (#462)
 * Rename `estimate_scr()` to `est_seroincidence_by()` (#439)
 * Rename `estimate_scr()` to `est_seroincidence()` (#432)
 * Rename argument `curve_params` to `sr_params` for estimation functions (#424)